+++ src/python/pants/backend/core/tasks/BUILD
@@ -5,17 +5,9 @@ target(
   name = 'all',
   dependencies = [
     ':builddictionary',
-    ':cloc',
     ':dependees',
-    ':filemap',
     ':filter',
-    ':list_owners',
-    ':listtargets',
-    ':minimal_cover',
-    ':pathdeps',
-    ':paths',
     ':scm_publish',
-    ':sorttargets',
   ]
 )
 
@@ -48,33 +40,11 @@ resources(
 )
 
 python_library(
-  name = 'cloc',
-  sources = ['cloc.py'],
-  dependencies = [
-    'src/python/pants/base:build_environment',
-    'src/python/pants/base:exceptions',
-    'src/python/pants/base:workunit',
-    'src/python/pants/binaries:binary_util',
-    'src/python/pants/task',
-    'src/python/pants/util:contextutil',
-  ],
-)
-
-python_library(
   name = 'dependees',
   sources = ['dependees.py'],
   dependencies = [
-    ':target_filter_task_mixin',
-    'src/python/pants/base:build_environment',
-    'src/python/pants/task',
-  ],
-)
-
-python_library(
-  name = 'filemap',
-  sources = ['filemap.py'],
-  dependencies = [
-    'src/python/pants/task',
+    'src/python/pants/backend/graph_info/tasks',
+    'src/python/pants/base:deprecated',
   ],
 )
 
@@ -82,48 +52,8 @@ python_library(
   name = 'filter',
   sources = ['filter.py'],
   dependencies = [
-    ':target_filter_task_mixin',
-    'src/python/pants/base:build_environment',
-    'src/python/pants/base:cmd_line_spec_parser',
-    'src/python/pants/base:exceptions',
-    'src/python/pants/build_graph',
-    'src/python/pants/task',
-    'src/python/pants/util:filtering',
-  ],
-)
-
-python_library(
-  name = 'listtargets',
-  sources = ['listtargets.py'],
-  dependencies = [
-    'src/python/pants/base:exceptions',
-    'src/python/pants/task',
-  ],
-)
-
-python_library(
-  name = 'minimal_cover',
-  sources = ['minimal_cover.py'],
-  dependencies = [
-    'src/python/pants/task',
-  ],
-)
-
-python_library(
-  name = 'pathdeps',
-  sources = ['pathdeps.py'],
-  dependencies = [
-    'src/python/pants/task',
-  ],
-)
-
-python_library(
-  name = 'paths',
-  sources = ['paths.py'],
-  dependencies = [
-    'src/python/pants/base:exceptions',
-    'src/python/pants/task',
-    'src/python/pants/util:strutil',
+    'src/python/pants/backend/graph_info/tasks',
+    'src/python/pants/base:deprecated',
   ],
 )
 
@@ -148,33 +78,6 @@ python_library(
   sources = ['scm_publish.py'],
   dependencies = [
     'src/python/pants/base:deprecated',
-    'src/python/pants/task',
-  ],
-)
-
-python_library(
-  name = 'sorttargets',
-  sources = ['sorttargets.py'],
-  dependencies = [
-    'src/python/pants/build_graph',
-    'src/python/pants/task',
-  ],
-)
-
-python_library(
-  name = 'target_filter_task_mixin',
-  sources = ['target_filter_task_mixin.py'],
-  dependencies = [
-    'src/python/pants/base:exceptions',
-    'src/python/pants/task',
-  ]
-)
-
-python_library(
-  name = 'list_owners',
-  sources = ['list_owners.py'],
-  dependencies = [
-    'src/python/pants/build_graph',
     'src/python/pants/task',
   ],
 )
