+++ src/python/pants/backend/graph_info/BUILD
@@ -1,0 +1,12 @@
+# Copyright 2015 Pants project contributors (see CONTRIBUTORS.md).
+# Licensed under the Apache License, Version 2.0 (see LICENSE).
+
+
+python_library(
+  name = 'plugin',
+  sources = ['register.py'],
+  dependencies = [
+    'src/python/pants/backend/graph_info/tasks',
+    'src/python/pants/goal:task_registrar',
+  ]
+)
