+++ src/python/pants/backend/graph_info/tasks/BUILD
@@ -1,0 +1,27 @@
+# Copyright 2014 Pants project contributors (see CONTRIBUTORS.md).
+# Licensed under the Apache License, Version 2.0 (see LICENSE).
+
+python_library(
+  name = 'tasks',
+  sources = globs('*.py'),
+  dependencies = [
+    '3rdparty/python:docutils',
+    '3rdparty/python:setuptools',
+    '3rdparty/python/twitter/commons:twitter.common.collections',
+    'src/python/pants/backend/python:python_requirements',
+    'src/python/pants/base:build_environment',
+    'src/python/pants/base:build_manual',
+    'src/python/pants/base:cmd_line_spec_parser',
+    'src/python/pants/base:exceptions',
+    'src/python/pants/base:generator',
+    'src/python/pants/base:workunit',
+    'src/python/pants/binaries:binary_util',
+    'src/python/pants/build_graph',
+    'src/python/pants/goal',
+    'src/python/pants/task',
+    'src/python/pants/util:contextutil',
+    'src/python/pants/util:dirutil',
+    'src/python/pants/util:filtering',
+    'src/python/pants/util:strutil',
+  ],
+)
