+++ src/python/pants/backend/graph_info/tasks/cloc.py
