+++ src/python/pants/backend/graph_info/tasks/minimal_cover.py
