+++ src/python/pants/backend/graph_info/tasks/paths.py
@@ -5,7 +5,6 @@
 from __future__ import (absolute_import, division, generators, nested_scopes, print_function,
                         unicode_literals, with_statement)
 
-import copy
 from collections import deque
 
 from pants.base.exceptions import TaskError
