+++ tests/python/pants_test/backend/core/tasks/BUILD
@@ -2,73 +2,13 @@
 # Licensed under the Apache License, Version 2.0 (see LICENSE).
 
 python_tests(
-  name='cache_cleanup',
-  sources=['test_cache_cleanup.py'],
-  dependencies=[
+  name = 'cache_cleanup_integration',
+  sources = ['test_cache_cleanup_integration.py'],
+  dependencies = [
     'src/python/pants/backend/jvm/tasks/jvm_compile:zinc',
     'src/python/pants/util:dirutil',
     'tests/python/pants_test:int-test',
     'src/python/pants/util:contextutil',
-  ]
-)
-
-python_tests(
-  name='cloc',
-  sources=['test_cloc.py'],
-  coverage=['pants.backend.core.tasks.cloc'],
-  dependencies=[
-    'src/python/pants/backend/core/tasks:cloc',
-    'src/python/pants/backend/jvm/targets:java',
-    'src/python/pants/backend/python/targets:python',
-    'src/python/pants/base:build_environment',
-    'tests/python/pants_test/tasks:task_test_base',
-  ]
-)
-
-python_tests(
-  name = 'dependees',
-  sources = ['test_dependees.py'],
-  dependencies = [
-    'src/python/pants/backend/codegen/targets:java',
-    'src/python/pants/backend/core/tasks:dependees',
-    'src/python/pants/backend/jvm/targets:java',
-    'src/python/pants/backend/jvm/targets:jvm',
-    'src/python/pants/backend/python/targets:python',
-    'src/python/pants/build_graph',
-    'tests/python/pants_test/tasks:task_test_base',
-  ]
-)
-
-python_tests(
-  name = 'filter',
-  sources = ['test_filter.py'],
-  dependencies = [
-    'src/python/pants/backend/core/tasks:filter',
-    'src/python/pants/backend/docgen/targets',
-    'src/python/pants/backend/jvm/targets:java',
-    'src/python/pants/backend/python/targets:python',
-    'src/python/pants/base:exceptions',
-    'src/python/pants/build_graph',
-    'tests/python/pants_test/tasks:task_test_base',
   ],
-)
-
-python_tests(
-  name='paths',
-  sources=['test_paths.py'],
-  dependencies=[
-    'src/python/pants/backend/core/tasks:paths',
-    'src/python/pants/base:exceptions',
-    'tests/python/pants_test/tasks:task_test_base',
-  ]
-)
-
-python_tests(
-  name='target_filter_task_mixin',
-  sources=['test_target_filter_task_mixin.py'],
-  dependencies=[
-    'src/python/pants/backend/core/tasks:target_filter_task_mixin',
-    'src/python/pants/build_graph',
-    'tests/python/pants_test/tasks:task_test_base',
-  ]
+  tags = {'integration'}
 )
