+++ tests/python/pants_test/backend/graph_info/tasks/test_filemap.py
@@ -8,7 +8,7 @@ from __future__ import (absolute_import, division, generators, nested_scopes, pr
 import os
 from textwrap import dedent
 
-from pants.backend.core.tasks.filemap import Filemap
+from pants.backend.graph_info.tasks.filemap import Filemap
 from pants.backend.python.targets.python_library import PythonLibrary
 from pants.build_graph.build_file_aliases import BuildFileAliases
 from pants_test.tasks.task_test_base import ConsoleTaskTestBase
