+++ tests/python/pants_test/backend/graph_info/tasks/test_filter.py
@@ -7,8 +7,8 @@ from __future__ import (absolute_import, division, generators, nested_scopes, pr
 
 from textwrap import dedent
 
-from pants.backend.core.tasks.filter import Filter
 from pants.backend.docgen.targets.doc import Page
+from pants.backend.graph_info.tasks.filter import Filter
 from pants.backend.jvm.targets.java_library import JavaLibrary
 from pants.backend.python.targets.python_library import PythonLibrary
 from pants.backend.python.targets.python_requirement_library import PythonRequirementLibrary
