+++ tests/python/pants_test/backend/graph_info/tasks/test_list_owners.py
@@ -7,7 +7,7 @@ from __future__ import (absolute_import, division, generators, nested_scopes, pr
 
 from textwrap import dedent
 
-from pants.backend.core.tasks.list_owners import ListOwners
+from pants.backend.graph_info.tasks.list_owners import ListOwners
 from pants.backend.python.targets.python_library import PythonLibrary
 from pants.base.exceptions import TaskError
 from pants.build_graph.build_file_aliases import BuildFileAliases
