+++ tests/python/pants_test/backend/graph_info/tasks/test_minimal_cover.py
@@ -7,7 +7,7 @@ from __future__ import (absolute_import, division, generators, nested_scopes, pr
 
 from textwrap import dedent
 
-from pants.backend.core.tasks.minimal_cover import MinimalCover
+from pants.backend.graph_info.tasks.minimal_cover import MinimalCover
 from pants.backend.python.targets.python_library import PythonLibrary
 from pants.build_graph.build_file_aliases import BuildFileAliases
 from pants_test.tasks.task_test_base import ConsoleTaskTestBase
