+++ tests/python/pants_test/backend/graph_info/tasks/test_sorttargets.py
@@ -7,7 +7,7 @@ from __future__ import (absolute_import, division, generators, nested_scopes, pr
 
 from textwrap import dedent
 
-from pants.backend.core.tasks.sorttargets import SortTargets
+from pants.backend.graph_info.tasks.sorttargets import SortTargets
 from pants.backend.python.targets.python_library import PythonLibrary
 from pants.build_graph.build_file_aliases import BuildFileAliases
 from pants_test.tasks.task_test_base import ConsoleTaskTestBase
