+++ tests/python/pants_test/tasks/BUILD
@@ -80,63 +80,6 @@ python_tests(
 )
 
 python_tests(
-  name = 'filemap',
-  sources = ['test_filemap.py'],
-  dependencies = [
-    ':task_test_base',
-    'src/python/pants/backend/core/tasks:filemap',
-    'src/python/pants/backend/python/targets:python',
-    'src/python/pants/build_graph',
-  ],
-)
-
-python_tests(
-  name = 'listtargets',
-  sources = ['test_listtargets.py'],
-  dependencies = [
-    ':task_test_base',
-    'src/python/pants/backend/core/tasks:listtargets',
-    'src/python/pants/backend/jvm:artifact',
-    'src/python/pants/backend/jvm/targets:java',
-    'src/python/pants/backend/jvm/targets:jvm',
-    'src/python/pants/build_graph',
-  ],
-)
-
-python_tests(
-  name = 'minimal_cover',
-  sources = ['test_minimal_cover.py'],
-  dependencies = [
-    ':task_test_base',
-    'src/python/pants/backend/core/tasks:minimal_cover',
-    'src/python/pants/backend/python/targets:python',
-    'src/python/pants/build_graph',
-  ],
-)
-
-python_tests(
-  name = 'sorttargets',
-  sources = ['test_sorttargets.py'],
-  dependencies = [
-    ':task_test_base',
-    'src/python/pants/backend/python/targets:python',
-    'src/python/pants/backend/core/tasks:sorttargets',
-    'src/python/pants/build_graph',
-  ],
-)
-
-python_tests(
-  name = 'list_owners',
-  sources = ['test_list_owners.py'],
-  dependencies = [
-    'src/python/pants/backend/core/tasks:list_owners',
-    'src/python/pants/backend/python/targets:python',
-    'src/python/pants/base:exceptions',
-    'src/python/pants/build_graph',
-  ],
-)
-
-python_tests(
   name = 'clean_all_integration',
   sources = ['test_clean_all_integration.py'],
   dependencies = [
